Reject PCO in the untyped MergeTree compression settings
Close the `allow_experimental_codecs` bypass and the late-failure of the untyped
MergeTree compression settings. `default_compression_codec`, `marks_compression_codec`
and `primary_key_compression_codec` are resolved without a column type, so they
never went through the column-codec experimental gate: `default_compression_codec
= 'PCO'` silently worked (re-resolved with the real type at write time), and the
marks/primary-key settings only failed at the first write.

`MergeTreeSettingsImpl::checkCompressionCodecSettings` (called from `sanityCheck`,
i.e. CREATE and ALTER MODIFY SETTING) now rejects an experimental codec in all
three settings and a codec that `requiresColumnTypeToCompress` in the two applied
to untyped streams. `CompressionCodecMultiple` already propagates both properties,
so a chain such as `PCO, ZSTD(1)` is rejected too. Extends `04513_pco_codec_gating`.

Also drops a redundant cast in `CompressionCodecPco.cpp` and an unused include in
`pco.h`.

+++ src/Compression/CompressionCodecPco.cpp
@@ -196,7 +196,7 @@ UInt32 CompressionCodecPco::doCompressData(const char * source, UInt32 source_si
     memcpy(&dest[2], source, bytes_to_skip);
 
     UInt64 n = (source_size - bytes_to_skip) / data_bytes_size;
-    UInt64 raw_payload_size = static_cast<UInt64>(n) * data_bytes_size;
+    UInt64 raw_payload_size = n * data_bytes_size;
     auto * payload = reinterpret_cast<uint8_t *>(dest) + 2 + bytes_to_skip;
 
     /// Give the encoder exactly the raw payload size as its budget. If the `.pco` stream would be
